>Fixed pipeline failure due to data type change in Tansaction class

+++ src/test/java/uk/co/asepstrath/bank/TransactionTest.java
@@ -18,7 +18,7 @@ public class TransactionTest {
     @BeforeEach
     public void init(){
         c = new Controller();
-        t = new Transaction("WID", "DID", "Timestamp", "ID", 400, "GBP");
+        t = new Transaction("WID", "DID", "Timestamp", "ID", BigDecimal.valueOf(400), "GBP");
         WithdrawAcc = new Account("WID", "Name", 2456.90, "GBP", "Type");
         DepositAcc = new Account("DID", "Name", 3553.09, "GBP", "Type");
